[Refactor] types 위치 변경 #3
